+++ payjoin-relay/src/main.rs
@@ -1,5 +1,6 @@
 use std::net::TcpListener;
 use std::thread::spawn;
+
 use tungstenite::accept;
 
 fn main() {
@@ -8,7 +9,7 @@ fn main() {
     let server = TcpListener::bind("127.0.0.1:3012").unwrap();
     println!("REElay listening on 127.0.0.1:3012 😡");
     for stream in server.incoming() {
-        spawn (move || {
+        spawn(move || {
             let mut websocket = accept(stream.unwrap()).unwrap();
             loop {
                 let msg = websocket.read_message().unwrap();
